requirements.txt
fastapi[all]==0.111.0
uvicorn[standard]==0.30.1

sqlalchemy==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0          # explicitly pin bcrypt to avoid surprises

pydantic==2.7.4
email-validator==2.1.1
python-multipart==0.0.9

redis==5.0.7
aioredis==2.0.1
websockets==12.0

python-dotenv==1.0.1
pillow==10.4.0